web_scraper.py
# web_scraper.py
import requests
from bs4 import BeautifulSoup
from typing import Dict, List
import time
from rich.progress import Progress, SpinnerColumn, TextColumn
from rich.console import Console

console = Console()

class PolicyScraper:
    """Scrape UK planning policies from official sources"""
    
    POLICY_SOURCES = {
        "NPPF": "https://www.gov.uk/government/publications/national-planning-policy-framework--2",
        "PPG": "https://www.gov.uk/government/collections/planning-practice-guidance",
        "PDR": "https://www.gov.uk/guidance/when-is-permission-required",
        "London Plan": "https://www.london.gov.uk/programmes-strategies/planning/london-plan"
    }
    
    def scrape_framework_content(self, framework: str, council: str = None) -> str:
        """Scrape content for a specific framework"""
        
        # For production, implement actual web scraping
        # This is mock content for demonstration
        mock_content = {
# web_scraper.py (continued)
            "NPPF": """
National Planning Policy Framework (2023)

Key Principles:
- Achieving sustainable development
- Plan-led system  
- Presumption in favour of sustainable development

Design Policies:
- Para 126: Good design is a key aspect of sustainable development
- Para 130: Planning decisions should ensure developments:
  a) Function well and add to overall quality of the area
  b) Are visually attractive with good architecture
  c) Are sympathetic to local character and history
  d) Optimize the potential of the site
  e) Create safe, inclusive and accessible places
  
- Para 134: Development that is not well designed should be refused
- Para 135: Local planning authorities should seek to ensure that the quality of approved development is not materially diminished between permission and completion

Residential Amenity:
- Para 130(f): Create places with a high standard of amenity for existing and future users
- Para 185: Ensure developments are appropriate for their location, avoiding noise and pollution impacts
            """,
            
            "PDR": """
Permitted Development Rights - Technical Guidance

Class A - Enlargement of a dwellinghouse:
- Single storey rear extensions: 6m (semi/terraced), 8m (detached)
- Two storey rear extensions: 3m for all house types
- Maximum height: 4m
- Maximum eaves height: 3m
- No extension beyond side elevation fronting highway
- Materials must be similar in appearance

Class B - Roof additions:
- Volume allowance: 40m³ (terraced), 50m³ (semi/detached)
- No extension beyond plane of existing roof slope fronting highway
- Maximum height not to exceed highest part of existing roof
- Minimum 20cm from eaves
- No verandas, balconies or raised platforms

Class AA - Upward extensions (dwellinghouses):
- Single storey only
- Must be detached house
- Maximum 3.5m additional height
- Engineering report required
- Prior approval needed
            """,
            
            "LDF": f"""
{council or 'Local'} Development Framework

Core Strategy Policies:
- CS1: Sustainable Development
- CS5: Design and Character
- CS7: Housing Mix and Density
- CS8: Open Space and Nature Conservation

Development Management Policies:
- DM1: Design Quality
  * Respect local character and context
  * High quality materials and detailing
  * Appropriate scale and massing
  
- DM2: Residential Extensions
  * Subordinate to main building
  * Respect building line
  * 45-degree rule for neighboring properties
  * Maintain adequate garden space
  
- DM10: Trees and Landscaping
  * Retain existing trees where possible
  * Replacement planting required
  * Landscape plans for major developments
            """,
            
            "LP": """
The London Plan 2021

Good Growth Objectives:
- GG1: Building strong and inclusive communities
- GG2: Making the best use of land
- GG3: Creating a healthy city

Design Policies:
- Policy D3 Optimising site capacity through the design-led approach
- Policy D4 Delivering good design
- Policy D5 Inclusive design
- Policy D6 Housing quality and standards
  * Minimum space standards
  * Private outdoor space requirements
  * Dual aspect requirements
  
- Policy D14 Noise
  * Agent of Change principle
  * Acoustic design statements

Housing Standards:
- Minimum 37 sqm (1b1p), 50 sqm (1b2p), 61 sqm (2b3p), 70 sqm (2b4p)
- Minimum ceiling height 2.5m for 75% of GIA
- Private outdoor space: 5 sqm (1-2 person), +1 sqm per additional person
            """,
            
            "BRE": """
Building Regulations Approved Documents

Part A - Structure
- Structural safety and stability requirements
- Foundations appropriate to ground conditions

Part B - Fire Safety
- Means of escape
- Fire resistance of structure
- Access for fire service

Part L - Conservation of fuel and power
- Energy efficiency standards
- U-values for walls, roofs, floors, windows
- Air permeability requirements

Part M - Access to and use of buildings
- Level access requirements
- Accessible WC provisions
- Circulation space standards

Part O - Overheating (new 2022)
- Overheating mitigation in new residential buildings
- Dynamic thermal modeling may be required
            """,
            
            "SPD": f"""
{council or 'Local'} Supplementary Planning Documents

Residential Design Guide SPD:
- Minimum rear garden depths: 10m
- Side space standards: 1m minimum to boundary
- Daylight/sunlight standards (BRE Guide)
- Privacy: 21m back-to-back, 12m back-to-side

Shopfront Design Guide SPD:
- Traditional proportions and materials
- Retention of architectural features
- Appropriate signage and lighting

Sustainable Design and Construction SPD:
- BREEAM standards for commercial
- Water efficiency measures
- Sustainable drainage systems (SuDS)
            """
        }
        
        # Simulate fetching with progress indicator
        with Progress(
            SpinnerColumn(),
            TextColumn("[progress.description]{task.description}"),
            console=console
        ) as progress:
            task = progress.add_task(f"Fetching {framework} content...", total=None)
            time.sleep(1)  # Simulate network delay
        
        content = mock_content.get(framework, f"Framework content for {framework}")
        console.print(f"[green]✓[/green] Retrieved {framework}")
        return content
    
    def get_council_policies(self, council: str, frameworks: List[str]) -> Dict[str, str]:
        """Get all selected frameworks content"""
        results = {}
        for framework in frameworks:
            results[framework] = self.scrape_framework_content(framework, council)
        return results