requirements.txt
# requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0
rich>=13.7.0
click>=8.1.7
PyPDF2>=3.0.0
inquirer>=3.1.0